PTSD-14 Feat : UserInfoContext 추가 개발
UserInfoContext 추가 개발 : time 계산 기능 추가

+++ hooks/contexts/user/UserInfoContext.tsx
@@ -3,20 +3,28 @@ import { apiTokenValidation } from '@hooks/apis/auth/common';
 import useCodeMsgBundle from '@hooks/utils/useCodeMsgBundle';
 import { useToastContext } from '@hooks/utils/useToastContext';
 import { deleteCookie, getCookie, setCookie } from 'cookies-next';
+import dayjs from 'dayjs';
 import jwtDecode from 'jwt-decode';
 import { useRouter } from 'next/router';
 import { createContext, useCallback, useEffect, useState } from 'react';
-import { useMutation } from 'react-query';
 
 interface UserInfoContext {
   userInfo?: UserInfoInterface;
   setInUserInfo?: (info: UserInfoInterface) => void;
-  handleTokenInfo?: () => void;
+  handleTokenInfo: () => void;
   deleteToken?: () => void;
   signOut?: () => void;
+  time?: (expTime: number) => {
+    minute: number;
+    seconds: number;
+  };
 }
 
-const UserInfoContext = createContext<UserInfoContext>({});
+const UserInfoContext = createContext<UserInfoContext>({
+  handleTokenInfo: () => {
+    return;
+  },
+});
 
 interface Props {
   children: JSX.Element | JSX.Element[];
@@ -26,8 +34,19 @@ const UserInfoProvider = ({ children }: Props): JSX.Element => {
   const router = useRouter();
   const toast = useToastContext();
   const msg = useCodeMsgBundle();
-  const { mutate: mutateTokenValidation } = useMutation(apiTokenValidation);
   const [userInfo, setUserInfo] = useState<UserInfoInterface>();
+
+  /**UserInfoContext 유효 시간 계산 */
+  const time = useCallback((expTime: number) => {
+    const now = dayjs();
+    const exp = expTime * 1000;
+    const lastDate = dayjs(exp);
+    const diffS = lastDate.diff(now, 's');
+    const minute = Math.floor(diffS / 60);
+    const m = minute * 60;
+    const seconds = diffS - m;
+    return { minute, seconds };
+  }, []);
 
   /**UserInfoContext 회원가입 라우팅 기능  */
   const signinPageRoute = useCallback(() => {
@@ -58,41 +77,45 @@ const UserInfoProvider = ({ children }: Props): JSX.Element => {
   }, []);
 
   /**UserInfoContext 토큰 리프레쉬 */
-  const handleTokenInfo = useCallback(() => {
+  const handleTokenInfo = useCallback(async () => {
     const token = getCookie('refreshToken');
-    const refreshToken = typeof token === 'string' ? token : '';
-    mutateTokenValidation(refreshToken, {
-      onSuccess(data, variables, context) {
-        if (data.data.code) {
-          if (data.data.code !== '0000') {
-            if (data.data.code === '0049') {
-              toast?.on(
-                '다른 기기에서 로그인 하였습니다. \n 로그인 페이지로 이동합니다.',
-                'info',
-              );
-              return;
-            } else {
-              toast?.on(msg.errMsg(data.data.code), 'error');
+    if (typeof document !== 'undefined') {
+      const refreshToken = typeof token === 'string' ? token : '';
+      if (refreshToken) {
+        await apiTokenValidation(refreshToken)
+          .then((data) => {
+            if (data.data.code) {
+              if (data.data.code !== '0000') {
+                if (data.data.code === '0049') {
+                  toast?.on(
+                    '다른 기기에서 로그인 하였습니다. \n 로그인 페이지로 이동합니다.',
+                    'info',
+                  );
+                  return;
+                } else {
+                  toast?.on(msg.errMsg(data.data.code), 'error');
+                }
+                deleteCookie('accessToken');
+                deleteCookie('refreshToken');
+                signOut();
+                return;
+              } else {
+                setCookie('accessToken', data.data.accessToken);
+                setCookie('refreshToken', data.data.refreshToken);
+              }
             }
-            deleteCookie('accessToken');
-            deleteCookie('refreshToken');
+          })
+          .catch(() => {
+            console.log('4');
+            toast?.on(
+              '토큰 유효기간이 만료 되었습니다. \n 로그인 페이지로 이동합니다.',
+              'error',
+            );
             signOut();
-            return;
-          } else {
-            setCookie('accessToken', data.data.accessToken);
-            setCookie('refreshToken', data.data.refreshToken);
-          }
-        }
-      },
-      onError(error, variables, context) {
-        toast?.on(
-          '토큰 유효기간이 만료 되었습니다. \n 로그인 페이지로 이동합니다.',
-          'error',
-        );
-        signOut();
-      },
-    });
-  }, [msg, mutateTokenValidation, signOut, toast]);
+          });
+      }
+    }
+  }, [msg, signOut, toast]);
 
   useEffect(() => {
     if (router.pathname !== '/signin') {
@@ -120,6 +143,7 @@ const UserInfoProvider = ({ children }: Props): JSX.Element => {
     handleTokenInfo,
     deleteToken,
     signOut,
+    time,
   };
 
   return (
